refactor(repositories): заменил метод get на find в репозиториях
META: #99905

+++ backend/src/User/SignIn/Command/DeleteToken.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\User\SignIn\Command;
 
 use App\Infrastructure\AsService;
+use App\User\Security\Service\TokenException;
 use App\User\User\Domain\UserTokenId;
 use App\User\User\Domain\UserTokenRepository;
 
@@ -18,7 +19,11 @@ final readonly class DeleteToken
 
     public function __invoke(UserTokenId $userTokenId): void
     {
-        $userToken = $this->userTokenRepository->getById($userTokenId);
+        $userToken = $this->userTokenRepository->findById($userTokenId);
+
+        if ($userToken === null) {
+            throw new TokenException('Токен не найден');
+        }
 
         $this->userTokenRepository->remove($userToken);
     }
